+++ jest.config.js
@@ -23,6 +23,9 @@ module.exports = {
     "^.+\\.[jt]sx?$": "babel-jest"
   },
   testPathIgnorePatterns: ["/node_modules/", "/dist/", "/template/"],
+  moduleNameMapper: {
+    "\\.svg": "<rootDir>/__mocks__/svgrMock.js"
+  },
   // Ref https://github.com/facebook/jest/issues/2070#issuecomment-431706685
   // Todo(steve): remove next line when issue fixed.
   modulePathIgnorePatterns: ["<rootDir>/packages/[^/]*/src/.*/__mocks__"],