+++ packages/brick-types/src/menu.ts
@@ -1,6 +1,6 @@
 import { LocationDescriptor } from "history";
 import { ThemeType } from "antd/lib/icon";
-import { IconProp } from "@fortawesome/fontawesome-svg-core";
+import { IconName, IconPrefix } from "@fortawesome/fontawesome-svg-core";
 
 export interface SidebarMenu {
   title: string;
@@ -33,7 +33,7 @@ export interface SidebarMenuGroup {
   key?: string;
 }
 
-export type MenuIcon = AntdIcon | FaIcon;
+export type MenuIcon = AntdIcon | FaIcon | EasyopsIcon;
 
 export interface AntdIcon {
   lib: "antd";
@@ -43,5 +43,12 @@ export interface AntdIcon {
 
 export interface FaIcon {
   lib: "fa";
-  icon: IconProp;
+  icon: IconName;
+  prefix?: IconPrefix;
+}
+
+export interface EasyopsIcon {
+  lib: "easyops";
+  icon: string;
+  category?: string;
 }