+++ dll/ace/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@0.1.5...@dll/ace@0.1.6) (2019-09-30)
+
+**Note:** Version bump only for package @dll/ace
+
+## [0.1.5](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@0.1.4...@dll/ace@0.1.5) (2019-09-29)
+
+**Note:** Version bump only for package @dll/ace
+
+## [0.1.4](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@0.1.3...@dll/ace@0.1.4) (2019-09-25)
+
+**Note:** Version bump only for package @dll/ace
+
+## [0.1.3](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@0.1.2...@dll/ace@0.1.3) (2019-09-24)
+
+**Note:** Version bump only for package @dll/ace
+
+## [0.1.2](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@0.1.1...@dll/ace@0.1.2) (2019-09-23)
+
+**Note:** Version bump only for package @dll/ace
+
 ## [0.1.1](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@0.1.0...@dll/ace@0.1.1) (2019-09-20)
 
 **Note:** Version bump only for package @dll/ace