+++ packages/brick-scripts/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.16](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-scripts@0.14.15...@easyops/brick-scripts@0.14.16) (2019-09-30)
+
+**Note:** Version bump only for package @easyops/brick-scripts
+
+## [0.14.15](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-scripts@0.14.14...@easyops/brick-scripts@0.14.15) (2019-09-29)
+
+**Note:** Version bump only for package @easyops/brick-scripts
+
+## [0.14.14](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-scripts@0.14.13...@easyops/brick-scripts@0.14.14) (2019-09-25)
+
+**Note:** Version bump only for package @easyops/brick-scripts
+
+## [0.14.13](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-scripts@0.14.12...@easyops/brick-scripts@0.14.13) (2019-09-24)
+
+**Note:** Version bump only for package @easyops/brick-scripts
+
+## [0.14.12](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-scripts@0.14.11...@easyops/brick-scripts@0.14.12) (2019-09-23)
+
+**Note:** Version bump only for package @easyops/brick-scripts
+
 ## [0.14.11](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-scripts@0.14.10...@easyops/brick-scripts@0.14.11) (2019-09-20)
 
 **Note:** Version bump only for package @easyops/brick-scripts